Major refactor to allow for scaling of capability. Can now handle boulder problems. Various bug fixes to tick analytics.
Refactored so that a full unique dataframe df_usend_uniq with scraped metrics is merged to the full user dataframe. User modifications such as grade homogenization are applied on df_usendm. Finally a filtered dataframe df_usendf is visualized.
Boulder problems are now handled in tandem with routes during the scrape and modification phase.
Fixed a bug in tick style type extraction that couldn't find "tr" ticks by removing the \s\s in the search regex.
Made it so that tick counts are properly scaled by pitches attempted in that tick.
Fixed bug that counted subsequent redpoints after the first as additional attempts.